modified Dmt Charge Commission in new format.

+++ setting-service/src/main/java/com/toshaniFintech/user_service/dto/request/DmtChargeCommissionRequest.java
@@ -5,22 +5,20 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class DmtChargeCommissionRequest {
     @NotNull(message = "Min amount is required")
     @DecimalMin(value = "0.0", inclusive = true, message = "Min amount must be greater than or equal to 0")
-    private BigDecimal minAmount;
+    private Double minAmount;
 
     @NotNull(message = "Max amount is required")
     @DecimalMin(value = "0.0", inclusive = true, message = "Max amount must be greater than or equal to 0")
-    private BigDecimal maxAmount;
+    private Double maxAmount;
 
     @NotBlank(message = "Commission type is required")
     private String commType;
 
     @NotNull(message = "Commission is required")
     @DecimalMin(value = "0.0", inclusive = true, message = "Commission must be greater than or equal to 0")
-    private BigDecimal comm;
+    private Double comm;
 }
