+++ setting-service/src/main/java/com/toshaniFintech/user_service/dto/response/DmtChargeCommissionResponse.java
@@ -8,10 +8,10 @@ import java.math.BigDecimal;
 public class DmtChargeCommissionResponse {
     private String id;
     private String slNo;
-    private BigDecimal minAmount;
-    private BigDecimal maxAmount;
+    private Double minAmount;
+    private Double maxAmount;
     private String commType;
-    private BigDecimal comm;
+    private Double comm;
     private boolean active;
 }
